Minor improvements for pretty good documentation.
It's already pretty good.

I caught a couple words in headings that should be capitalized and a consistently missing comma in one location. The biggest change (in terms of the number of words shuffled around) was making sentences be active rather than passive (and consequently shorter and thus faster and easier to read).