Speed up binding by minimizing the sources that are used
Previously, as each property in hierarchy was being bound, every
source would be checked for a possible value. This was inefficient
when there was a large number of property sources and a particular
hierarchy of properties only existed in a small number of those
sources.

This commit generalizes context's existing support for binding with
one specific source (as used for lists) to support binding with n
specific sources. For each property that's bound, the relevant
sources are found by asking if they contain the current property
or a descendent of it. Those that do not are not used for the binding
of the property and all of its descendents.