Make annotation attribute values more concise
When the value of an annotation's attribute references something
defined by a type that's contained by the annotation, the contained
type should be imported and referenced directly rather than via the
containing type. Consider the following:

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)

The repetition of SpringBootTest in both the annotation itself and
the value of one of its attributes is overly verbose. Instead, we
prefer to import WebEnvironment and refer to it directly:

@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)