Specify platform when exporting an image's layers
This change is overly broad at the moment. It raises the minimum
Docker API version to 1.48 for anyone building an image with a
buildpack that is not part of the builder. It needs to be narrowed
down so that only those specifying an image plaform that's
not the host's platform are affected.